requirements.txt
asgiref==3.6.0
Django==3.2.5
djangorestframework==3.12.4
pytz==2023.3
sqlparse==0.4.4

django-extensions==3.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
django-import-export==2.7.1
drf-yasg==1.20.0
django-crispy-forms==1.12.0
django-bootstrap-breadcrumbs==0.9.2
pydantic==1.9.0
python-environ==0.4.54
